+++ src/main/java/com/desk/java/apiclient/service/UserService.java
@@ -74,7 +74,7 @@ public interface UserService {
      * @return a user
      */
     @GET(USERS_URI + "/{id}")
-    Call<User> getUser(@Path("id") int userId);
+    Call<User> getUser(@Path("id") long userId);
 
     /**
      * List all of the user's mobile devices.
@@ -84,7 +84,7 @@ public interface UserService {
      * @return a mobile device api response
      */
     @GET(USERS_URI + "/{id}/" + MOBILE_DEVICES_URI)
-    Call<ApiResponse<MobileDevice>> getMobileDevicesForUser(@Path("id") int userId);
+    Call<ApiResponse<MobileDevice>> getMobileDevicesForUser(@Path("id") long userId);
 
     /**
      * Creates a mobile device for the current user
@@ -102,7 +102,7 @@ public interface UserService {
      * @return nothing
      */
     @DELETE(USERS_URI + "/current/" + MOBILE_DEVICES_URI + "/{id}")
-    Call<Void> deleteMobileDevice(@Path("id") int id);
+    Call<Void> deleteMobileDevice(@Path("id") long id);
 
     /**
      * Retrieve a list of mobile device settings.
@@ -113,7 +113,7 @@ public interface UserService {
      * @return a setting api response
      */
     @GET(USERS_URI + "/{userId}/" + MOBILE_DEVICES_URI + "/{deviceId}/" + SETTINGS_URI)
-    Call<ApiResponse<Setting>> getMobileDevicesSettings(@Path("userId") int userId, @Path("deviceId") int deviceId);
+    Call<ApiResponse<Setting>> getMobileDevicesSettings(@Path("userId") long userId, @Path("deviceId") long deviceId);
 
     /**
      * Update a mobile device setting
@@ -126,6 +126,6 @@ public interface UserService {
      * @return a setting
      */
     @PATCH(USERS_URI + "/{userId}/" + MOBILE_DEVICES_URI + "/{deviceId}/" + SETTINGS_URI + "/{settingId}")
-    Call<Setting> updateMobileDeviceSetting(@Path("userId") int userId, @Path("deviceId") int deviceId,
-                                   @Path("settingId") int settingId, @Body SettingUpdate update);
+    Call<Setting> updateMobileDeviceSetting(@Path("userId") long userId, @Path("deviceId") long deviceId,
+                                   @Path("settingId") long settingId, @Body SettingUpdate update);
 }
