+++ src/main/java/com/salesforce/scmt/desk/DeskContactMigration.java
@@ -59,18 +59,18 @@ public class DeskContactMigration<D extends Serializable> extends DeskBase<D>
 	private static final int DESK_PAGE_SIZE_CUSTOMER = 100;
 
     @Override
-    protected int getId(D d)
+    protected long getId(D d)
     {
         Customer c = (Customer) d;
         return c.getId();
     }
 
     @Override
-    protected int getUpdatedAt(D d)
+    protected long getUpdatedAt(D d)
     {
         try {
             Customer c = (Customer) d;
-            return (int) (ISO8601Utils.parse(c.getUpdatedAt(), new ParsePosition(0)).getTime() / 1000);
+            return (long) (ISO8601Utils.parse(c.getUpdatedAt(), new ParsePosition(0)).getTime() / 1000);
         } catch (ParseException e) {
             e.printStackTrace();
         }
